# swagger/swagger.yaml
basePath: /api/v0
definitions:
...
  consoleapi.SwaggerErrorResponse:
    properties:
      error:
        example: request validation failed
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for StorX Monitor server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: StorX Monitor API
  version: "1.0"
paths:
...
  /auth/account:
    delete:
      consumes:
      - application/json
      description: '**Full route:** `DELETE /api/v0/auth/account`'
      parameters:
      - description: Target email and verification password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.DeleteAuthAccountSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete account (admin)
      tags:
      - auth-account
    get:
      description: '**Full route:** `GET /api/v0/auth/account`'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.AuthAccountSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Get current account
      tags:
      - auth-account
    patch:
      consumes:
      - application/json
      description: '**Full route:** `PATCH /api/v0/auth/account`'
      parameters:
      - description: Name fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.UpdateAuthAccountSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Update account display name
      tags:
      - auth-account
  /auth/account/change-password:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/account/change-password`'
      parameters:
      - description: From GET /config when CSRF enabled
        in: header
        name: X-CSRF-Token
        type: string
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.AuthChangePasswordSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Change password (logged in, password already set)
      tags:
      - auth-account
...
  /auth/account/set-password:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/account/set-password` — requires
        logged-in session (`_tokenKey` cookie) and CSRF when enabled. **Swagger:**
        Authorize **CookieAuth** with session token value; Authorize **CSRFAuth**
        with `csrfToken` from `GET /config`; then call this route. Body: `{ "newPassword":
        "..." }` (8–64 chars). Returns 409 if password already set.'
      parameters:
      - description: New password only
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.AuthSetPasswordSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password set
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      - CSRFAuth: []
      summary: Set initial password (optional after Google / email verify)
      tags:
      - auth-set-password
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/forgot-password`'
      parameters:
      - description: Email and captcha
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.AuthForgotPasswordSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email sent if account exists
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      summary: Request password reset email
      tags:
      - auth-password-recovery
...
  /auth/mfa/disable:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/mfa/disable`'
      parameters:
      - description: Passcode or recovery code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.MFADisableSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Disable MFA
      tags:
      - auth-account
  /auth/mfa/enable:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/mfa/enable`'
      parameters:
      - description: TOTP passcode
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.MFAEnableSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Recovery codes
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Enable MFA
      tags:
      - auth-account
  /auth/mfa/generate-recovery-codes:
    post:
      description: '**Full route:** `POST /api/v0/auth/mfa/generate-recovery-codes`'
      produces:
      - application/json
      responses:
        "200":
          description: Recovery codes
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Generate MFA recovery codes
      tags:
      - auth-account
  /auth/mfa/generate-secret-key:
    post:
      description: '**Full route:** `POST /api/v0/auth/mfa/generate-secret-key`'
      produces:
      - application/json
      responses:
        "200":
          description: TOTP secret key string
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Generate MFA secret key
      tags:
      - auth-account
  /auth/mfa/regenerate-recovery-codes:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/mfa/regenerate-recovery-codes`'
      parameters:
      - description: Passcode or recovery code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.MFARegenerateRecoveryCodesSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New recovery codes
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Regenerate MFA recovery codes
      tags:
      - auth-account
  /auth/refresh-session:
    post:
      description: '**Full route:** `POST /api/v0/auth/refresh-session`'
      produces:
      - application/json
      responses:
        "200":
          description: New session expiresAt (RFC3339)
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Refresh login session
      tags:
      - auth-account
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/reset-password`'
      parameters:
      - description: Recovery token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.AuthResetPasswordSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset; session cookie cleared
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/consoleapi.AuthResetPasswordMFARequiredResponse'
      summary: Reset password with recovery token
      tags:
      - auth-password-recovery
  /auth/token:
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/auth/token`'
      parameters:
      - description: Email, password, captcha
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.AuthTokenSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.AuthCredentialLoginSwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CSRFAuth: []
      summary: Email + password login
      tags:
      - auth-email-login
...
  /config:
    get:
      description: '**Full route:** `GET /api/v0/config` — public bootstrap (no auth).
        Call before `POST /auth/token`: read `captcha.login` (hCaptcha/reCAPTCHA site
        keys) and `csrfToken` when CSRF is enabled; send `captchaResponse` and optional
        `X-CSRF-Token` on login. Also returns `apiBaseURL`, feature flags, Stripe
        key, and other `FrontendConfig` fields.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleweb.FrontendConfig'
        "500":
          description: Internal Server Error
      summary: Get frontend satellite configuration
      tags:
      - auth-email-login
      - config
...
  /seller/auth/token:
    post:
      consumes:
      - application/json
      description: '**Route:** `POST /api/v0/seller/auth/token`. Sets `_seller_tokenKey`
        session cookie.'
      parameters:
      - description: Email, password, captcha
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/seller.SellerAuthTokenSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/seller.SellerAuthTokenSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/seller.SellerAuthErrorResponse'
      security:
      - SellerCSRFAuth: []
      summary: Reseller email + password login
      tags:
      - reseller-auth
  /seller/branding:
    get:
      description: '**Route:** `GET /api/v0/seller/branding`. Logo/favicon fields
        are public asset URLs.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/seller.ResellerBrandingConfig'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/seller.SellerAuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/seller.SellerAuthErrorResponse'
      security:
      - SellerCookieAuth: []
      summary: Get reseller branding
      tags:
      - reseller-branding
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer / API key (rare for console routes).
    in: header
    name: Authorization
    type: apiKey
  CSRFAuth:
    description: 'When `console.csrf-protection-enabled` is true: copy `csrfToken`
      from `GET /config` and Authorize here. Must match `csrf_token` cookie (set by
      GET /config or synced by Swagger UI). Required for `POST /auth/token`, `POST
      /auth/account/set-password`, and other CSRF-protected routes.'
    in: header
    name: X-CSRF-Token
    type: apiKey
  CookieAuth:
    description: Session cookie after login. **Swagger Authorize:** paste the token
      value only (from `POST /auth/token` response `token` field). Do not prefix with
      `Cookie:` or `_tokenKey=`. Copied curl uses `-b '_tokenKey=...'` (not a header).
      Login via Swagger auto-fills this.
    in: cookie
    name: _tokenKey
    type: apiKey
  SellerCSRFAuth:
    description: 'When `seller.csrf-protection-enabled` is true: copy `csrfToken`
      from `GET /seller/config` and Authorize here. Must match `csrf_token` cookie
      (set by GET /seller/config). Required for `POST /seller/auth/token`, MFA routes,
      password change, and other CSRF-protected seller routes.'
    in: header
    name: X-CSRF-Token
    type: apiKey
  SellerCookieAuth:
    description: Session cookie after reseller Google login. **Swagger Authorize:**
      paste the token value only (from `GET /seller/auth/google` response `token`
      field). Seller peer cookie name is `_seller_tokenKey` (not console `_tokenKey`).
    in: cookie
    name: _seller_tokenKey
    type: apiKey
swagger: "2.0"
...